<Upload book image and Generate PDF file and Download it>

+++ controllers/admin.js
@@ -1,8 +1,10 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
 
+const fileHelper = require('../helper/file');
+
 exports.getAddProduct = (req, res, next) => {
-    res.render('admin/edit-product',{
+  res.render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
     editing: false,
@@ -13,9 +15,9 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
-  const errors = validationResult(req);
-  if(!errors.isEmpty()){
-    return res.status(422).render('admin/edit-product',{
+  const image = req.file;
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/edit-product',
       editing: false,
@@ -24,34 +26,58 @@ exports.postAddProduct = (req, res, next) => {
         title: req.body.title,
         price: req.body.price,
         description: req.body.description,
-        imageUrl: req.body.imageUrl,
-      }, 
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: [],
+    });
+  }
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: req.body.title,
+        price: req.body.price,
+        description: req.body.description,
+      },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
+  const imageUrl = image.path;
+
   new Product({
     title: req.body.title,
     description: req.body.description,
     price: req.body.price,
-    imageUrl: 'https://cdn.pixabay.com/photo/2016/03/31/20/51/book-1296045_960_720.png',
+    imageUrl: imageUrl,
     userId: req.session.user,
   }).save()
-  .then(() =>{
-    res.redirect('/');
-  })
-  .catch(error => {
-    console.error('Error creating product:', error);
-    next(error); // Pass the error to the error-handling middleware
-  }); 
+    .then(() => {
+      res.redirect('/');
+    })
+    .catch(error => {
+      console.error('Error creating product:', error);
+      next(error); // Pass the error to the error-handling middleware
+    });
 }
 
 exports.deleteProduct = (req, res, next) => {
-  Product.deleteOne({
-    _id: req.params.productId,
-     userId: req.session.user._id
+  Product.findById(req.body.productId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl, next);
+      return Product.deleteOne({
+        _id: req.body.productId,
+        userId: req.session.user._id
+      })
     })
-    .then(result=>{
+    .then(result => {
       console.log('Product deleted:', result);
       res.redirect('/admin/products');
     })
@@ -59,6 +85,7 @@ exports.deleteProduct = (req, res, next) => {
       console.log(error);
       next(error);
     });
+
 }
 
 exports.getEditProduct = (req, res, next) => {
@@ -66,16 +93,16 @@ exports.getEditProduct = (req, res, next) => {
   if (!editMode) {
     return res.redirect('/');
   }
- Product.findById(req.params.productId)
-    .then(product=>{
+  Product.findById(req.params.productId)
+    .then(product => {
       if (!product) {
         return res.redirect('/');
       }
-      res.render('admin/edit-product',{
+      res.render('admin/edit-product', {
         pageTitle: 'Edit Product',
         path: '/admin/edit-product',
         editing: editMode,
-        product: product, 
+        product: product,
         hasError: false,
         errorMessage: null,
         validationErrors: [],
@@ -90,9 +117,10 @@ exports.getEditProduct = (req, res, next) => {
 };
 
 exports.postEditProduct = (req, res, next) => {
+  const image = req.file;
   const errors = validationResult(req);
-  if(!errors.isEmpty()){
-    return res.status(422).render('admin/edit-product',{
+  if (!errors.isEmpty()) {
+    return res.status(422).render('admin/edit-product', {
       pageTitle: 'Eddit Product',
       path: '/admin/edit-product',
       editing: true,
@@ -101,45 +129,47 @@ exports.postEditProduct = (req, res, next) => {
         title: req.body.title,
         price: req.body.price,
         description: req.body.description,
-        imageUrl: req.body.imageUrl,
         _id: req.body.productId
-      }, 
+      },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
   Product.findById(req.body.productId)
-  .then(product => {
-    if (product.userId.toString() !== req.session.user._id.toString()) {
-      return res.redirect('/');
-    }
-    product.title = req.body.title;
-    product.price = req.body.price;
-    product.description = req.body.description;
-    product.imageUrl = req.body.imageUrl;
-    return product.save() 
-    .then(result => {
-      console.log('Product updated:', result);
-      res.redirect('/');
+    .then(product => {
+      if (product.userId.toString() !== req.session.user._id.toString()) {
+        return res.redirect('/');
+      }
+      product.title = req.body.title;
+      product.price = req.body.price;
+      product.description = req.body.description;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      return product.save()
+        .then(result => {
+          console.log('Product updated:', result);
+          res.redirect('/');
+        })
     })
-  })
-  .catch(error => {
-    console.log(error);
-    next(error);
-  });
+    .catch(error => {
+      console.log(error);
+      next(error);
+    });
 }
 
-exports.getProducts = (req,res,next) =>{
-    Product.find({userId: req.session.user._id})
-      .then(products => {
-        res.render('admin/products', {
-          prods: products,
-          pageTitle: 'Admin Products',
-          path: '/admin/products',
-        });
-      })
-      .catch(error => {
-        console.log(error);
-        next(error);
+exports.getProducts = (req, res, next) => {
+  Product.find({ userId: req.session.user._id })
+    .then(products => {
+      res.render('admin/products', {
+        prods: products,
+        pageTitle: 'Admin Products',
+        path: '/admin/products',
       });
+    })
+    .catch(error => {
+      console.log(error);
+      next(error);
+    });
 };
